+++ src/settings/MappingProfiles/initialDetails/index.js
@@ -24,6 +24,11 @@ const initialValues = {
   MARC_AUTHORITY,
 };
 
+const KEEP_SUBFIELDS_INITIAL = {
+  [FOLIO_RECORD_TYPES.INVOICE.type]: ['acqUnitIds'],
+  [FOLIO_RECORD_TYPES.HOLDINGS.type]: ['permanentLocationId'],
+};
+
 const modifyInvoiceLinesFieldDetails = field => {
   const invoiceLineSubfield = { ...field.subfields[0] };
 
@@ -98,15 +103,13 @@ export const getInitialDetails = (entity, stripRepeatableFields = false) => {
       if (field.name === 'invoiceLines') {
         return modifyInvoiceLinesFieldDetails(field);
       }
-
-      if (field.name === 'acqUnitIds') {
-        return field;
-      }
     }
+
+    const keepSubfieldsInitial = KEEP_SUBFIELDS_INITIAL[entity]?.some(fieldName => fieldName === field.name);
 
     return {
       ...field,
-      subfields: [],
+      subfields: keepSubfieldsInitial ? field.subfields : [],
     };
   });